Prepare for odd-number context token exchanges
GSS mechanisms can support odd numbers of context tokens, in which case
the last context token is sent by the initiator to the acceptor, but if
the acceptor fails to accept it then it should produce one more context
token for the initiator, which then must process it with
gss_process_context_token() to extract usable error information.

Note that there's an ambiguity in the spec as to
gss_process_context_token(), and so we can't quite know if GSS_S_FAILURE
from it means "failure while processing the token" or "this is what
broke on the acceptor side".

In theory the Kerberos mechanism w/o mutual authentication can yield
this situation.  In practice no implementations send error tokens in
that case.  However, we're working on an extension that allows for more
than two context tokens in the Kerberos mechanism, and anyways, knc
should be more mechanism-generic.

Finally, there's a very unfortunate bug in Heimdal and MIT Kerberos
whereby their gss_process_context_token() actually deletes the security
context when it returns GSS_S_COMPLETE, which means we must leak the
context rather than delete it in that case, at least for now.